pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aind-physiology-fip"
description = "A repository for Fiber Photometry data acquisition"
authors = [
    { name = "Bruno Cruz", email = "bruno.cruz@alleninstitute.org" },
]
license = {text = "MIT"}
requires-python = ">=3.11"
dynamic = ["version", "readme"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]

dependencies = [
  "aind_behavior_services",
]

[project.urls]
Documentation = "https://allenneuraldynamics.github.io/Aind.Physiology.Fip/"
Repository = "https://github.com/AllenNeuralDynamics/Aind.Physiology.Fip/"
Issues = "https://github.com/AllenNeuralDynamics/Aind.Physiology.Fip/issues"
Changelog = "https://github.com/AllenNeuralDynamics/Aind.Physiology.Fip/releases"

[project.optional-dependencies]

launcher = ["aind_behavior_experiment_launcher[aind-services]"]

dev = [
    "aind_physiology_fip[launcher]",
    "ruff",
    "codespell"
]

docs = [
    'Sphinx<7.3',
    'furo',
    'sphinx-jinja',
    'autodoc_pydantic[erdantic]',
    'sphinx-jsonschema',
    'sphinx-copybutton',
    "sphinx_mdinclude",
]

[project.scripts]
clabe = "aind_physiology_fip.launcher:main"
regenerate = "aind_physiology_fip.regenerate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aind_physiology_fip = ["py.typed"]

[tool.setuptools.dynamic]
version = {attr = "aind_physiology_fip.__version__"}
readme = {file = "README.md", content-type = "text/markdown"}

[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'google' }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build,uv.lock'
ignore-words-list = 'nd'
